@@ -4,7 +4,7 @@ version: 0.2.0
 homepage: https://appblaze.io
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.20.0"
 
 dependencies:
